Add test for Api Gw
